fmt: Allow downcasting a FmtSubscriber to its components
This commit overrides the `downcast_raw` method on `FmtSubscriber` so
that a `FmtSubscriber` can also be downcast to pointers to its generic
components (its on-event callback, filter, and `NewVisitor`).

This will allow implementations of the components of a `FmtSubscriber`
to support "out-of-band" APIs that don't require knowledge of the types
of the _other_ components to downcast the subscriber to a concrete type.
This is intended for use-cases such as dynamic filter reloading (#42).

Signed-off-by: Eliza Weisman <eliza@buoyant.io>